+++ django_request_formatter/exceptions.py
@@ -1,0 +1,4 @@
+
+class RequestValidationError(Exception):
+    def __init__(self, message, code=None, params=None):
+        super().__init__(message, code, params)