+++ tests/testapp/forms.py
@@ -1,3 +1,5 @@
+from django.core.exceptions import ValidationError
+
 from django_request_formatter.forms import Form
 from django_request_formatter import fields
 
@@ -18,4 +20,12 @@ class AlbumForm(Form):
     year = fields.IntegerField()
     artist = fields.FormField(form=ArtistForm)
     songs = fields.FormFieldList(form=SongForm)
-    created_at = fields.DateField()
+    metadata = fields.DictionaryField(fields.DateTimeField())
+
+    def validate_year(self, value):
+        if value == "1992":
+            raise ValidationError("Year 1992 is forbidden!")
+
+    def validate(self):
+        if (self._data['year'] == "1998") and (self._data['artist'] == "Nirvana"):
+            raise ValidationError("Sounds like a bullshit")